Replace static routes with template finder
