+++ requirements.txt
@@ -1,6 +1,7 @@
 canonicalwebteam.flask-base==0.7.3
 canonicalwebteam.http==1.0.3
 canonicalwebteam.blog==6.3.1
+canonicalwebteam.templatefinder==1.0.0
 canonicalwebteam.yaml_responses[flask]==1.1.1
 python-dateutil==2.8.1
 pyyaml==5.3.1
