+++ webapp/app.py
@@ -6,11 +6,11 @@ A Flask application for jp.ubuntu.com
 import yaml
 import flask
 import talisker
+import webapp.template_utils as template_utils
+
 from canonicalwebteam.blog import build_blueprint, BlogViews, BlogAPI
 from canonicalwebteam.flask_base.app import FlaskBase
-
-from webapp.blueprint import jp_website
-import webapp.template_utils as template_utils
+from canonicalwebteam.templatefinder import TemplateFinder
 
 session = talisker.requests.get_session()
 app = FlaskBase(
@@ -18,9 +18,10 @@ app = FlaskBase(
     "jp.ubuntu.com",
     template_folder="../templates",
     static_folder="../static",
+    template_404="404.html",
+    template_500="500.html",
 )
 
-app.register_blueprint(jp_website)
 blog_views = BlogViews(
     api=BlogAPI(
         session=session,
@@ -39,16 +40,6 @@ with open("releases.yaml") as releases:
     releases = yaml.load(releases, Loader=yaml.FullLoader)
 
 
-@app.errorhandler(404)
-def not_found_error(error):
-    return flask.render_template("404.html"), 404
-
-
-@app.errorhandler(500)
-def internal_error(error):
-    return flask.render_template("500.html"), 500
-
-
 # Template context
 @app.context_processor
 def context():
@@ -63,6 +54,25 @@ def context():
         "architecture": flask.request.args.get("architecture", ""),
     }
 
+
+@app.route("/favicon.ico")
+def favicon():
+    return flask.redirect(
+        "https://res.cloudinary.com/canonical/image/fetch/q_auto,f_auto/"
+        "https://assets.ubuntu.com/v1/088fd1bf-favicon.ico"
+    )
+
+
+@app.route("/robots.txt")
+def robots():
+    return flask.Response("", mimetype="text/plain")
+
+
+# All other routes
+template_finder_view = TemplateFinder.as_view("template_finder")
+template_finder_view._exclude_xframe_options_header = True
+app.add_url_rule("/", view_func=template_finder_view)
+app.add_url_rule("/<path:subpath>", view_func=template_finder_view)
 
 if __name__ == "__main__":
     app.run(host="0.0.0.0")
